backend/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asttokens==2.4.1
async-timeout==5.0.1
asyncpg==0.30.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.7
fonttools==4.55.6
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
ipykernel==6.29.4
ipython==8.23.0
jedi==0.19.1
joblib==1.4.2
jupyter_client==8.6.1
jupyter_core==5.7.2
jwt==1.3.1
kiwisolver==1.4.8
matplotlib==3.10.0
matplotlib-inline==0.1.6
nest-asyncio==1.6.0
neurokit2==0.2.10
numpy==2.2.2
opencv-python-headless==4.11.0.86
packaging==24.0
pandas==2.2.3
parso==0.8.4
passlib==1.7.4
pillow==11.1.0
platformdirs==4.2.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2==2.9.10
psycopg2-binary==2.9.10
pure-eval==0.2.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.17.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==25.1.2
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.37
sqlmodel==0.0.22
stack-data==0.6.3
starlette==0.45.2
threadpoolctl==3.5.0
tornado==6.4
traitlets==5.14.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.4
wcwidth==0.2.13
websockets==14.2
